rename mgs_manager to mgs
